refactor(future-cli): split out generatOutputData helper
