+++ future-cli.js
@@ -55,7 +55,9 @@ const Blob = require(blobPath).Blob
 
 const OpenJsCad = require(openjscadPath).OpenJsCad
 const openscad = require(openscadPath)
+
 const makeInputFormatHandlers = require('./js/utils/inputFormatHandlers')
+const generateOutputData = require('./js/utils/generateOutputData')
 
 const meta = {
   producer: `OpenJSCAD ${OpenJsCad.version}`,
@@ -98,7 +100,7 @@ let src = fs.readFileSync(inputFile, inputFile.match(/\.stl$/i) ? 'binary' : 'UT
 src = inputFormatHandlers[inputFormat](src, inputFile, outputFile)
 
 // -- convert from JSCAD script into the desired output format
-const outputData = generateOutputData(outputFormat, src, gMainParam)
+const outputData = generateOutputData(outputFormat, src, gMainParam, meta, CSG, CAG, lib, Blob)
 // -- and write it to disk
 writeOutputDataToFile(outputFile, outputData)
 
@@ -224,52 +226,6 @@ function evaluateSource (src, helperFunctions, mainParams) {
     }
     csg = ouput
   }
-}
-
-function generateOutputData (outputFormat, src, gMainParam) {
-  let data
-  if (outputFormat === 'jscad' || outputFormat === 'js') {
-    data = new Blob([src], { type: 'application/javascript' })
-  } else {
-    const scad = fs.readFileSync(path.resolve(lib, './openscad.js'))
-    // console.log("render jscad to "+outputFormat)
-    // console.log(JSON.stringify(gMainParam))
-    const fullSrc = `${src}\n${scad}\nmain(_getParameterDefinitions(${JSON.stringify(gMainParam)}))\n`
-    let csg = eval(fullSrc) // src + '\n' + scad + '\nmain(_getParameterDefinitions(' + JSON.stringify(gMainParam) + '))\n') // *.jscad + openscad.js + main()
-
-    if (csg.length) {
-      var o = csg[0]
-      if (o instanceof CAG) {
-        o = o.extrude({offset: [0, 0, 0.1]})
-      }
-      for (let i = 1; i < csg.length; i++) {
-        var c = csg[i]
-        if (c instanceof CAG) {
-          c = c.extrude({offset: [0, 0, 0.1]})
-        }
-        o = o.unionForNonIntersecting(c)
-      }
-      csg = o
-    }
-
-    const outputFormatHandlers = {
-      'amf': () => csg.toAMFString(meta), // CSG to AMF
-      'stlb': () => csg.toStlBinary(), // CSG to STL BINARY
-      'stl': () => csg.toStlString(), // CSG to STL ASCII
-      'stla': () => csg.toStlString(), // CSG to STL ASCII
-      'dxf': () => cag.toDxf(), // CAG to DXF
-      'svg': () => cag.toSvg(), // CAG to SVG
-      'x3d': () => csg.toX3D(), // CSG to X3D Only possible via browsers
-      'json': () => csg.toJSON(), // CSG or CAG to JSON
-      undefined: () => {
-        console.log('ERROR: only jscad, stl, amf, dxf, svg or json as output format')
-        process.exit(1)
-      }
-    }
-    data = outputFormatHandlers[outputFormat]()
-  }
-  console.log(`Blob: type [${data.type}] size [${data.size}]`)
-  return data
 }
 
 function writeOutputDataToFile (outputFile, outputData) {
